Framegraph: add renderpasses image and buffer accesses
Track which images and buffers are being read/written by each
renderpass.

The code defines new structures for images and buffer informations: it
would be tempting to directly reuse the Vulkan API
objects (e.g. pointers to ImageObjectʳ), but note that these objects
might well be modified during the whole capture mutation. Also, this
resource information will eventually be represented as a protobuf
type, so the new types we define here act as drafts for those protobuf
types.

Bug: b/172218645
Test: manual, on several traces.